requirements.txt
easyocr
pyuac
pynput
Pillow
numpy
pywin32